requirements.txt
animals.py==1.0.0
beautifulsoup4==4.11.1
cowpy==1.1.5
dateparser==1.1.1
discord.py==1.7.3
emoji==1.7.0
fuzzywuzzy==0.18.0
gibberish==0.4.0
gTTS==2.2.3
nltk==3.6.5
numpy==1.19.5
opencv_python==4.5.1.48
pandas
psutil==5.9.1
PyDictionary==2.0.1
pypokedex==1.6.0
pytimeparse==1.1.8
PyYAML==3.13
randomname==0.1.5
requests==2.26.0
suntime==1.2.5
validators==0.20.0
wikipedia==1.4.0
wiktionaryparser==0.0.97
youtube_dl==2021.6.6
